125B Rev 4 PCB
Updated hardware design for the 125B PCB to include hardware muting.  This allows the signal to be muted during true bypass relay switching, which reduces any popping that might happen otherwise.

Updated the example software to use the hardware mute when using the bypass.  Also, added basic menu support to the OLED screen with some settings for the Tremolo and hardware device.