Updated 125B hardware build instructions.
The Bournes PRS11R pots have become almost impossible to find these days, so I updated the built instructions and the enclosure drill template to accommidate using Alpha pots instead. These stand a little taller off the PCB so the enclosure and the build instructions had to be slightly modified to accommidate for the differences.